refactor(web): break the api/state/Toast import cycle
Move authTokenSignal into auth.js and the toast signals plus addToast and
removeToast into toasts.js, both leaf modules. api.js imports only those
two. state.js and Toast.js re-export the moved names so every existing
import keeps working. No behavior change.
